Add files via upload
ros_img_save.py currently handles mission directory creation and publishes the mission name (date time of start of mission) for gobi_subscriber and bag_record to subscribe to, and use for writing files to the correct places.  May want to let bag_record figure out the mission directory and publish and have the two cameras subscribe, get the mission name, and then go write files when "record" topic is True.  subprocess may not be the best way to start rosbag in bag_record.py.